Decouples zipkin thrift interfaces and top-levels Dependency view
Finagle/Scrooge do not natively support multiple thrift interfaces on
the same port, eventhough other libraries do (such as FB Swift). To work
around this, zipkin used to expose query and collector services via
a thrift service hierarchy. This coupling was accidental and limits our
ability to be modular.

This change breaks up the thrift interfaces, particularly around the
dependency aggregates, addressing the Finagle limitation in code as
opposed to in thrift definitions.

Since we only expose a single aggregate function (dependencies), and
we very often have questions about it, this replaces the abstract term
aggregates with dependency. It also adds mysql instructions which
actually work!