Clarifies contract of DependencyStore and backfills tests
This standardizing on micros for dependency start/end interval and
backfills tests in a way that doesn't flake on timezones. It also
clarifies docs and converges the DependencySource/Sink as prior to our
move to json.

In the near future, we are going to migrate from Thrift to Json api.
This backfills tests and revises the contract around dependencies so
that future work isn't confusing or difficult to implement portably.

Dependencies implementations were confusing, using a combination of
microseconds, dates, milliseconds, Time, Duration and Calendar
instances to do the job. There were also incorrect docs, confusing
which of startTime or endTime was the default anchor of the interval.

This was very difficult to understand and easy to break, as there were
no tests that cover, for example, what the result value should be. It
was also difficult to understand the difference between the Dependencies
api and limitations of the Cassandra implementation.