API doc improvements
* changed version to 2
* Added example values for span. This makes the example JSON in the
spans POST valid. (Fixes #46).

The spans POST documentation is still a bit misleading. With the
parameter named "spans", it seems to imply that it's expecting

```json
{
    "spans": [{<...span...>}, {<....span...>}]
}
```

but it actually just wants:

```json
[{<...span...>}, {<...span...>}]
```

I couldn't figure out a way to acurately represent this in Swagger 2. It
might be possible using [Swagger
3 request body descriptions.](https://swagger.io/docs/specification/describing-request-body/)