Reverts to dummy response instead of a compile dep on google's empty
Empty is the only thing that requires a compile dependency on google
(not standard proto) types. For example, if you use square wire
compiler, you need to add a google library now to compile our proto.

Since we haven't really released the service, yet, and empty is binary
compatible with an equally empty response, we can change.

This reverts to what @ewhauser had originally, which makes things
more portable.