Adds reverse-mapped proto file for the json scheme in OpenApi/Swagger
The reverse engineers the json format from our swagger into proto in
attempts to unblock envoy from being able to generate json. It appears
the only allowed tool is protobuf. Usually, we wouldn't host a file
like this, but the otherwise is no-one being able to move off json v1
unless they go to protobuf, and we know that is not likely to occur.

See https://github.com/envoyproxy/envoy/pull/6985